Add page 3 to view the data inputed in page 1 & 2
Deploy additional feature of viewing the form data from page 1 and
table data from page 2. The entered data can be viewed in table
format for both the form data and table data. Deploy this feature
to view the data from the frontend directly instead of going to
database to check if data is getting inputed properly.